requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
coverage==7.3.2
cron-descriptor==1.4.0
cykhash==2.0.1
Cython==3.0.6
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.0
django-nose==1.4.7
django-request-logging==0.7.5
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
fiona==1.9.5
geopandas==0.14.1
googlemaps==4.10.0
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
json-fix==0.5.2
kombu==5.3.3
MarkupSafe==2.1.3
mysqlclient==2.2.0
nose==1.3.7
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pika==1.3.2
prompt-toolkit==3.0.41
PyJWT==2.8.0
pymongo==4.6.0
pyproj==3.6.1
pyrobuf==0.9.3
pyrosm==0.6.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-rapidjson==1.13
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
